chore: Rename Dialog.tsx to ReviewDialog.tsx
